Refactor Azure operations into separate class
